projects_with_fastapi/TodoApp/backend_sqllite/src/requirements.txt
fastapi
pydantic
email-validator 
uvicorn 
SQLAlchemy
passlib
bcrypt
pyjwt
python-multipart
asyncpg==0.21.0
databases[postgresql]==0.2.6

# dev 
pytest==5.3.2
requests==2.22.0